Don't push agenda layout when changing timezone
